requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.26.0
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
psutil==7.1.3
